Move `file.rs` to `utils`
Moves the `src/file.rs` to `src/utils` and changes the namespace in all
affected files.
