+++ src/bin/zwreec.rs
@@ -145,7 +145,7 @@ fn main() {
     // only for testing
     debug!("(1) {}", zwreec::frontend::temp_hello());
     debug!("(2) {}", zwreec::backend::temp_hello());
-    debug!("(3) {}", zwreec::file::temp_hello());
+    debug!("(3) {}", zwreec::utils::file::temp_hello());
 
     info!("main finished");
 }
